irqchip: renesas-intswcd: fix quirk definition
Commit e61448414aa6 ("irqchip: renesas-intswcd: add quirk to handle
edge interrupt as level" implemented a non-working quirk definition.

That was developed under impression that client interrupt definition
in device tree can have arbitrary length and actual length is passed to
irqchip's xlate method. But that is not true: xlate method always gets
length defined by #interrupt-cells in the irqchip's node, and longer
client interrupt definition is treated as definition of multiple
interrupts.

This commit reworks the quirk parser to use the intswcd node:

&intswcd {
	quirks {
		intp35 {
			interrupt = <R8A779F0_INTSWCD_INTP35>;
			force-level-low;
			setup = <0xdfed4b0c 0x4>;
		};
		...
	};
};

Property "setup" is a hack to configure the glitch filter. It defines
an address and a value; parser writes the value to the register at that
address. To be replaced with support inside pinctrl system.

Signed-off-by: Nikita Yushchenko <nikita.yoush@cogentembedded.com>