net: phy: attach reset suppliers early
Before this patch, phy was looking up it's reset supplier (either gpio
or other reset control) in mdiobus_register(). That code was executed
only if mdio device being registered was a phy, and was dependent
on of_node attached to phy's struct device (or perhaps some other way
to make getting gpio or reset control for device functional).

Problem is that in some cases, phy must be put out of reset earlier than
it is registered, namely when phy ids are being probed.

Previous patches added phy_device_initialize() and phy_device_probe_ids()
functions that implement unresetting phy when ids are probed. But those
needed reset suppliers to be attached early.

This patch implements that, by:

- moving management of reset suppliers from mdiobus code into explicit
  functions - mdio_device_attach_suppliers() and
  mdio_device_release_suppliers(),

- creating of_phy_device_initialize() - a wrapper that calls
  phy_device_initialize() and populates the returned phydev with
  information from device tree (which previously was done in
  of_mdiobus_phy_device_register() - the later function is removed)
  and also sets phydev's of_node and calls suppliers attach routine,

- converting all in-tree paths that set phydev's of_node to use
  of_phy_device_initialize() and phy_device_probe_ids().

Signed-off-by: Nikita Yushchenko <nikita.yoush@cogentembedded.com>